added dedicated functions and classes
