+++ src/config.ini
@@ -1,0 +1,8 @@
+[viz]
+ny_times_covid_date = 2020-03-30
+findthemasks_url = http://findthemasks.com/data.json
+request_headers = {"User-Agent": "Mozilla/5.0 (Windows NT 6.1) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/41.0.2228.0 Safari/537.3"}
+county_fips_download_url = https://github.com/ShyamW/Geocoding_Suite/blob/master/Lat_Lng_to_County_Data/county_Fips.txt
+geojson_url = https://raw.githubusercontent.com/plotly/datasets/master/geojson-counties-fips.json
+ny_times_county_data_url = https://github.com/nytimes/covid-19-data/raw/master/us-counties.csv
+hospital_download_url = https://docs.google.com/spreadsheet/ccc?key=15gZsozGQp-wdJaSngvLV13iCf_2mm2IsZpHOPxZtvtI&output=csv
